refactor: 장기판이 장기판 초기화된 기물 배치를 주입받도록 변경
