+++ src/test/java/boardSetting/PieceTest/HorseTest.java
@@ -1,9 +1,9 @@
-package janggiBoard.PieceTest;
+package boardSetting.PieceTest;
 
 import domain.position.Position;
 import domain.Team;
 import domain.piece.Blank;
-import domain.piece.Car;
+import domain.piece.Horse;
 import domain.piece.Piece;
 import domain.PieceProvider;
 import org.junit.jupiter.api.BeforeEach;
@@ -14,40 +14,37 @@ import java.util.Map;
 
 import static org.assertj.core.api.Assertions.*;
 
-public class CarTest {
-
-    private Car car;
+public class HorseTest {
     private TestPieceProvider testBoard;
+    private Horse horse;
 
     @BeforeEach
-    public void setUp() {
-        car = new Car(Team.CHO);
+    void setUp() {
+        horse = new Horse(Team.CHO);
         testBoard = new TestPieceProvider();
     }
 
     @Test
-    void 차는_목적지에_갈_수_있다() {
+    void 마가_목적지에_갈_수_있다() {
         Position currentPosition = new Position(5, 5);
-        Position targetPosition = new Position(7, 5);
+        Position targetPosition = new Position(3, 4);
 
         testBoard.setAllBlank();
-        boolean isCarMove = car.canMove(currentPosition, targetPosition, testBoard);
-        assertThat(isCarMove).isTrue();
+        boolean isCanMove = horse.canMove(currentPosition, targetPosition, testBoard);
+        assertThat(isCanMove).isTrue();
     }
 
     @Test
-    void 차는_목적지에_갈_수_없다() {
+    void 마가_목적지에_갈_수_없다() {
         Position currentPosition = new Position(5, 5);
-        Position targetPosition = new Position(7, 5);
+        Position targetPosition = new Position(3, 4);
 
         testBoard.setAllBlank();
-        testBoard.setBlank(new Position(6, 5));
+        testBoard.setBlank(new Position(4, 5));
 
-        boolean isCarMove = car.canMove(currentPosition, targetPosition, testBoard);
-
-        assertThat(isCarMove).isFalse();
+        boolean isCanMove = horse.canMove(currentPosition, targetPosition, testBoard);
+        assertThat(isCanMove).isFalse();
     }
-
 
     private static class TestPieceProvider implements PieceProvider {
         private final Map<Position, Boolean> boardState = new HashMap<>();
@@ -64,6 +61,11 @@ public class CarTest {
         @Override
         public boolean isBlank(Position position) {
             return boardState.getOrDefault(position, defaultState);
+        }
+
+        @Override
+        public boolean isCannon(Position position) {
+            return false;
         }
 
         @Override
