+++ src/test/java/boardSetting/PieceTest/PawnTest.java
@@ -1,0 +1,4 @@
+package boardSetting.PieceTest;
+
+public class PawnTest {
+}
