+++ src/test/java/boardSetting/strategyTest/CarStrategyTest.java
@@ -1,4 +1,4 @@
-package janggiBoard.strategyTest;
+package boardSetting.strategyTest;
 
 import domain.position.Position;
 import domain.piece.Blank;
@@ -70,6 +70,11 @@ public class CarStrategyTest {
         @Override
         public boolean isBlank(Position position) {
             return boardState.getOrDefault(position, defaultState);
+        }
+
+        @Override
+        public boolean isCannon(Position position) {
+            return false;
         }
 
         @Override
