--- src/test/java/janggiBoard/PieceTest/PawnTest.java
@@ -1,4 +1,0 @@
-package janggiBoard.PieceTest;
-
-public class PawnTest {
-}
